Shade in BouncyCastle
Because the container private keys are in a format that cannot be read by the
JVM, a limited portion of Bouncy Castle must be used.  This change shades that
portion of Bouncy Castle into the JAR relocating it so that it won't be
interfere with other implementations included in applications.