Incorporate CA Certs
Previously, an attempt at incorporating JRE CA certs didn't handle the
initialization of an SSL context in the way that Apache HTTPClient did it.
This meant that an NPE was thrown and the security provider was discarded.
This change implements a more robust version of this incorporation which
should be more resilient to various SSL context initialization strategies.

[cloudfoundry/java-buildpack#429]