Revert to Delegation
Previously, there was a lot of logic as the security provider tried to
replicate loading the standard JRE cacerts.  It also discarded any
initialization that happened with a non-empty KeyStore.  Needless to say this
was always going to cause bugs in the future because the behavior was so
custom that no one could predict its behavior with actually looking at the
code.  This change reverts much of that behavior to a delegate decorator. This
decorator wraps both "standard" key or trust store with the highest priority
and a container-based key or trust store with lowest priority.  This should be
much easier to predict when problems arise.