Cache Key and Trust Managers
Previously the Key and Trust Manager Factories would create new system and
file watching managers whenever they were asked for their artifacts.  This was
based on the idea that most frameworks only ask for each of these types once
and cache them for future usage.  It turns out that certain frameworks, and
certain usage patterns of other frameworks call for these artifacts multiple
times.  Creating new artifacts for these calls isn't horrible, but isn't as
efficient as it could be and prints enough log statements to give a user a
sense of unease.

This change updates the security provider to cache the system managers until
they are re-initialized and the file watching managers indefinitely.

[resolves cloudfoundry/java-buildpack#459]