Manager Properties
Previously, the addition of the SecurityProvider to an application meant that
you always got both the KeyManager and TrustManager for in that application
(although if certain artifacts were missing, these might be no ops).  There
are certain circumstances where you might want one or the other of the
managers though, so being able to disable them individually is useful.  This
change adds an inspection of the org.cloudfoundry.security.keymanager.enabled
and org.cloudfoundry.security.trustmanager.enabled properties to determine if
they should be disabled.  These properties default to true if unspecified.

[cloudfoundry/java-buildpack#552]